+++ app/src/main/kotlin/com/adesso/movee/internal/injection/module/NetworkModule.kt
@@ -3,17 +3,13 @@ package com.adesso.movee.internal.injection.module
 import android.content.Context
 import com.adesso.movee.BuildConfig
 import com.adesso.movee.data.remote.api.MovieService
-import com.adesso.movee.internal.util.DateAdapter
-import com.adesso.movee.internal.util.ImageJsonAdapter
 import com.adesso.movee.internal.util.NetworkStateHolder
 import com.adesso.movee.internal.util.api.ApiKeyInterceptor
 import com.adesso.movee.internal.util.api.ErrorHandlingInterceptor
 import com.adesso.movee.internal.util.api.RetryAfterInterceptor
 import com.chuckerteam.chucker.api.ChuckerInterceptor
 import com.moczul.ok2curl.CurlInterceptor
-import com.serjltt.moshi.adapters.Wrapped
 import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Lazy
 import dagger.Module
 import dagger.Provides
@@ -93,16 +89,5 @@ internal class NetworkModule {
     @Singleton
     fun provideMovieService(retrofit: Retrofit): MovieService {
         return retrofit.create(MovieService::class.java)
-    }
-
-    @Provides
-    @Singleton
-    fun provideMoshi(): Moshi {
-        return Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .add(ImageJsonAdapter())
-            .add(DateAdapter())
-            .add(Wrapped.ADAPTER_FACTORY)
-            .build()
     }
 }
